[bugfix] Texture coordinate must be float values

+++ Source/Core/GPU/Mesh.h
@@ -16,7 +16,7 @@ struct VertexFormat
 	VertexFormat(glm::vec3 position, 
 				glm::vec3 color = glm::vec3(1),
 				glm::vec3 normal = glm::vec3(0, 1, 0),
-				glm::uvec2 text_coord = glm::uvec2(0))
+				glm::vec2 text_coord = glm::vec2(0))
 		: position(position), normal(normal), color(color), text_coord(text_coord) { };
 
 	// position of the vertex
@@ -26,7 +26,7 @@ struct VertexFormat
 	glm::vec3 normal;
 
 	// vertex texture coordinate
-	glm::uvec2 text_coord;
+	glm::vec2 text_coord;
 
 	// vertex color
 	glm::vec3 color;
